.resources/.staging/python/cli/extract/github-stars-by-topic/requirements.txt
PyGithub
requests
Markdown
scikit-learn
bs4
dotenv